refactor: 지도 컴포넌트 재사용할 수 있게 리팩토링
